Add in basic hcard parsing functionality.
Given a URI find and return as a dict the first vcard in the URI.

When parsing if a property has an href in the top element, use that
as the value of the propery. If the propery is photo, use the src
or the img. Otherwise build the text of the subnodes of the property.

This is likely not entirely correct...